requirements.txt
pandas
matplotlib
h5py
scikit-learn
pycollimator
jax
pillow
openpyxl